robit now logs a date upon startup
The goal for this is that it sets up a framework for information
to be presented on bot startup. Ideally this could be expanded upon.

In addition to that, I would consider working to create a logging thread
somehow that keeps the file open. That would make things a whole lot
easier.